starts jest testing configs
