+++ src/mocha/suite/extension.oldtest.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-undef */
 import * as assert from 'assert';
 
 // You can import and use all API from the 'vscode' module
@@ -6,10 +7,10 @@ import * as vscode from 'vscode';
 // import * as myExtension from '../../extension';
 
 suite('Extension Test Suite', () => {
-	vscode.window.showInformationMessage('Start all tests.');
+  vscode.window.showInformationMessage('Start all tests.');
 
-	test('Sample test', () => {
-		assert.equal(-1, [1, 2, 3].indexOf(5));
-		assert.equal(-1, [1, 2, 3].indexOf(0));
-	});
+  test('Sample test', () => {
+    assert.equal(-1, [1, 2, 3].indexOf(5));
+    assert.equal(-1, [1, 2, 3].indexOf(0));
+  });
 });
