+++ src/test/sum.test.ts
@@ -1,0 +1,6 @@
+/* eslint-disable no-undef */
+import sum from './sum';
+
+test('adds 1 + 2 to equal 3', () => {
+  expect(sum(1, 2)).toBe(3);
+});
