+++ src/test/sum.ts
@@ -1,0 +1,3 @@
+export default function sum(a: number, b: number) {
+  return a + b;
+}
